Move CalendarDataResolver to data & DI
