+++ .gitignore
@@ -4,3 +4,4 @@ result
 emulator_config.json
 config.py
 .mypy_cache
+.localenvrc
